refactor/cleanup auth URI handling + errors
